Exploring the effect of deploying an outdoor small cell on system performance. Introduction: In this
In this project, I have designed a python project simulation which will simulate the downlink
behaviour of two base stations near a road and park/square communicating with mobiles
travelling the road and moving in the square park. The first base station will be an
omnidirectional macro cell located south of the centre of a 6 km section of road. The second
small cell will be an omnidirectional small cell located just north of the centre of the road in
a large open area park or town square.